requirements.txt
svg.path
PyMuPDF
Pillow
svgwrite
numpy
